so much work for a wip banner lol Add WIP (Work in Progress) Banner to the main app page
A new component, WipBanner, has been added to the App component. This banner takes place of the Header and Main components and signifies that the site is under construction. The NavBar CSS was tweaked to align with the new look of the site. The WipBanner comes with its own css module, which includes animations for the included Vite and React logos.